refactor: extracts some further properties from essential service module
