+++ embrace-android-sdk/config/detekt/baseline.xml
@@ -2,7 +2,7 @@
 <SmellBaseline>
   <ManuallySuppressedIssues></ManuallySuppressedIssues>
   <CurrentIssues>
-    <ID>LongParameterList:EssentialServiceModuleSupplier.kt$( initModule: InitModule, openTelemetryModule: OpenTelemetryModule, coreModule: CoreModule, workerThreadModule: WorkerThreadModule, systemServiceModule: SystemServiceModule, androidServicesModule: AndroidServicesModule, storageModule: StorageModule, customAppId: String?, logModuleProvider: Provider&lt;LogModule&gt;, featureModuleProvider: Provider&lt;FeatureModule&gt;, framework: AppFramework, configServiceProvider: (framework: AppFramework) -&gt; ConfigService? )</ID>
+    <ID>LongParameterList:EssentialServiceModuleSupplier.kt$( initModule: InitModule, openTelemetryModule: OpenTelemetryModule, coreModule: CoreModule, workerThreadModule: WorkerThreadModule, systemServiceModule: SystemServiceModule, androidServicesModule: AndroidServicesModule, storageModule: StorageModule, customAppId: String?, featureModuleProvider: Provider&lt;FeatureModule&gt;, framework: AppFramework, configServiceProvider: (framework: AppFramework) -&gt; ConfigService? )</ID>
     <ID>LongParameterList:SessionOrchestrationModuleSupplier.kt$( initModule: InitModule, openTelemetryModule: OpenTelemetryModule, androidServicesModule: AndroidServicesModule, essentialServiceModule: EssentialServiceModule, deliveryModule: DeliveryModule, workerThreadModule: WorkerThreadModule, dataSourceModule: DataSourceModule, payloadSourceModule: PayloadSourceModule, dataCaptureServiceModule: DataCaptureServiceModule, momentsModule: MomentsModule, logModule: LogModule )</ID>
     <ID>MaxLineLength:FakeModuleInitBootstrapper.kt$sessionOrchestrationModuleSupplier: SessionOrchestrationModuleSupplier = { _, _, _, _, _, _, _, _, _, _, _ -&gt; FakeSessionOrchestrationModule() }</ID>
     <ID>MaximumLineLength:FakeModuleInitBootstrapper.kt$ </ID>
