+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/injection/SessionOrchestrationModuleImpl.kt
@@ -1,5 +1,9 @@
 package io.embrace.android.embracesdk.internal.injection
 
+import io.embrace.android.embracesdk.internal.gating.EmbraceGatingService
+import io.embrace.android.embracesdk.internal.gating.GatingService
+import io.embrace.android.embracesdk.internal.session.EmbraceMemoryCleanerService
+import io.embrace.android.embracesdk.internal.session.MemoryCleanerService
 import io.embrace.android.embracesdk.internal.session.caching.PeriodicBackgroundActivityCacher
 import io.embrace.android.embracesdk.internal.session.caching.PeriodicSessionCacher
 import io.embrace.android.embracesdk.internal.session.message.PayloadFactory
@@ -26,9 +30,21 @@ internal class SessionOrchestrationModuleImpl(
     logModule: LogModule
 ) : SessionOrchestrationModule {
 
+    override val gatingService: GatingService by singleton {
+        EmbraceGatingService(
+            essentialServiceModule.configService,
+            logModule.logService,
+            initModule.logger
+        )
+    }
+
+    override val memoryCleanerService: MemoryCleanerService by singleton {
+        EmbraceMemoryCleanerService(logger = initModule.logger)
+    }
+
     override val payloadMessageCollatorImpl: PayloadMessageCollatorImpl by singleton {
         PayloadMessageCollatorImpl(
-            essentialServiceModule.gatingService,
+            gatingService,
             payloadSourceModule.sessionEnvelopeSource,
             androidServicesModule.preferencesService,
             openTelemetryModule.currentSessionSpan
@@ -60,7 +76,7 @@ internal class SessionOrchestrationModuleImpl(
 
     private val boundaryDelegate by singleton {
         OrchestratorBoundaryDelegate(
-            essentialServiceModule.memoryCleanerService,
+            memoryCleanerService,
             essentialServiceModule.userService,
             essentialServiceModule.sessionPropertiesService,
             essentialServiceModule.networkConnectivityService
