+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeEssentialServiceModule.kt
@@ -6,12 +6,9 @@ import io.embrace.android.embracesdk.fakes.FakeApiService
 import io.embrace.android.embracesdk.fakes.FakeApiUrlBuilder
 import io.embrace.android.embracesdk.fakes.FakeConfigService
 import io.embrace.android.embracesdk.fakes.FakeCpuInfoDelegate
-import io.embrace.android.embracesdk.fakes.FakeGatingService
 import io.embrace.android.embracesdk.fakes.FakeLogWriter
-import io.embrace.android.embracesdk.fakes.FakeMemoryCleanerService
 import io.embrace.android.embracesdk.fakes.FakePendingApiCallsSender
 import io.embrace.android.embracesdk.fakes.FakeProcessStateService
-import io.embrace.android.embracesdk.fakes.FakeRnBundleIdTracker
 import io.embrace.android.embracesdk.fakes.FakeSessionIdTracker
 import io.embrace.android.embracesdk.fakes.FakeSessionPropertiesService
 import io.embrace.android.embracesdk.fakes.FakeUserService
@@ -26,10 +23,8 @@ import io.embrace.android.embracesdk.internal.comms.api.ApiService
 import io.embrace.android.embracesdk.internal.comms.api.ApiUrlBuilder
 import io.embrace.android.embracesdk.internal.comms.delivery.PendingApiCallsSender
 import io.embrace.android.embracesdk.internal.config.ConfigService
-import io.embrace.android.embracesdk.internal.gating.GatingService
 import io.embrace.android.embracesdk.internal.injection.EssentialServiceModule
 import io.embrace.android.embracesdk.internal.logging.EmbLoggerImpl
-import io.embrace.android.embracesdk.internal.session.MemoryCleanerService
 import io.embrace.android.embracesdk.internal.session.id.SessionIdTracker
 import io.embrace.android.embracesdk.internal.session.lifecycle.ActivityTracker
 import io.embrace.android.embracesdk.internal.session.lifecycle.ProcessStateService
@@ -39,8 +34,6 @@ internal class FakeEssentialServiceModule(
     override val activityLifecycleTracker: ActivityTracker = FakeActivityTracker(),
     override val sessionIdTracker: SessionIdTracker = FakeSessionIdTracker(),
     override val configService: ConfigService = FakeConfigService(),
-    override val memoryCleanerService: MemoryCleanerService = FakeMemoryCleanerService(),
-    override val gatingService: GatingService = FakeGatingService(),
     override val apiClient: ApiClient = FakeApiClient(),
     override val userService: UserService = FakeUserService(),
     override val sharedObjectLoader: SharedObjectLoader = SharedObjectLoader(EmbLoggerImpl()),
@@ -51,5 +44,4 @@ internal class FakeEssentialServiceModule(
     override val logWriter: LogWriter = FakeLogWriter(),
     override val cpuInfoDelegate: CpuInfoDelegate = FakeCpuInfoDelegate(),
     override val sessionPropertiesService: FakeSessionPropertiesService = FakeSessionPropertiesService(),
-    override val rnBundleIdTracker: FakeRnBundleIdTracker = FakeRnBundleIdTracker()
 ) : EssentialServiceModule
