+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeSessionOrchestrationModule.kt
@@ -1,9 +1,13 @@
 package io.embrace.android.embracesdk.fakes.injection
 
+import io.embrace.android.embracesdk.fakes.FakeGatingService
+import io.embrace.android.embracesdk.fakes.FakeMemoryCleanerService
 import io.embrace.android.embracesdk.fakes.FakePayloadFactory
 import io.embrace.android.embracesdk.fakes.FakeSessionOrchestrator
 import io.embrace.android.embracesdk.fakes.FakeSessionSpanAttrPopulator
+import io.embrace.android.embracesdk.internal.gating.GatingService
 import io.embrace.android.embracesdk.internal.injection.SessionOrchestrationModule
+import io.embrace.android.embracesdk.internal.session.MemoryCleanerService
 import io.embrace.android.embracesdk.internal.session.caching.PeriodicBackgroundActivityCacher
 import io.embrace.android.embracesdk.internal.session.caching.PeriodicSessionCacher
 import io.embrace.android.embracesdk.internal.session.message.PayloadFactory
@@ -14,7 +18,9 @@ import io.embrace.android.embracesdk.internal.session.orchestrator.SessionSpanAt
 internal class FakeSessionOrchestrationModule(
     override val payloadFactory: PayloadFactory = FakePayloadFactory(),
     override val sessionOrchestrator: SessionOrchestrator = FakeSessionOrchestrator(),
-    override val sessionSpanAttrPopulator: SessionSpanAttrPopulator = FakeSessionSpanAttrPopulator()
+    override val sessionSpanAttrPopulator: SessionSpanAttrPopulator = FakeSessionSpanAttrPopulator(),
+    override val memoryCleanerService: MemoryCleanerService = FakeMemoryCleanerService(),
+    override val gatingService: GatingService = FakeGatingService(),
 ) : SessionOrchestrationModule {
 
     override val payloadMessageCollatorImpl: PayloadMessageCollatorImpl
